rl-botics/requirements.txt
absl-py==0.7.0
actionlib==1.11.13
alabaster==0.7.12
angles==1.9.11
astor==0.7.1
atari-py==0.1.7
Babel==2.6.0
bondpy==1.8.3
box2d-py==2.3.8
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.14
catkin-pkg==0.4.10
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
controller-manager==0.13.3
controller-manager-msgs==0.13.3
cv-bridge==1.12.8
cycler==0.10.0
Cython==0.29.6
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.5.50
empy==3.3.2
future==0.17.1
gast==0.2.2
gazebo-plugins==2.5.18
gazebo-ros==2.5.18
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
glfw==1.7.1
grpcio==1.19.0
# Editable Git install with no remote (gym==0.12.0)
-e /home/pal/Downloads/gym-master
gym-pomdp==0.0.1
h5py==2.9.0
idna==2.8
image-geometry==1.12.8
imageio==2.5.0
imagesize==1.1.0
interactive-markers==1.11.4
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
laser-geometry==1.6.4
lockfile==0.12.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
message-filters==1.12.14
mock==2.0.0
moveit-ros-planning-interface==0.9.15
moveit-ros-visualization==0.9.15
numpy==1.16.2
packaging==19.0
pbr==5.1.3
Pillow==5.4.1
pluginlib==1.11.3
protobuf==3.7.0
pycparser==2.19
pygame==1.9.4
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
pyparsing==2.3.1
PyQt5==5.12
PyQt5-sip==4.19.14
python-dateutil==2.8.0
python-qt-binding==0.3.4
pytz==2018.9
PyYAML==3.13
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.21.0
resource-retriever==1.12.4
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.7
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.16
scipy==1.2.1
sensor-msgs==1.12.7
six==1.12.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.0
srdfdom==0.4.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
urdfdom-py==0.3.3
urllib3==1.24.1
Werkzeug==0.14.1
xacro==1.11.3
